backend/modules/workspace/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
aiosqlite==0.22.1
alembic==1.13.1
psycopg2-binary==2.9.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
ulid-py==1.1.0

# Storage
boto3==1.34.14

# Development
black==23.12.1
ruff==0.1.13
mypy==1.8.0
